Issue #2871642 switch to factory creation of DrupalClient.
