--- fb_instant_articles.services.yml
@@ -1,4 +1,0 @@
-services:
-  fb_instant_articles.drupal_client:
-    class:  Drupal\fb_instant_articles\DrupalClient
-    arguments: ['@config.factory']
